Internal/2023.2/staging

+++ Tests/SRPTests/Projects/MultipleSRP_Tests/Assets/Tests/EditMode/Preview/AssetPreviewTests.cs
@@ -14,22 +14,39 @@ using Object = UnityEngine.Object;
 
 namespace Preview
 {
-    [TestFixture]
     public class AssetPreviewTests
     {
-        const string k_PreviewFolderPrefix = "Assets/Tests/EditMode/Preview";
+        public const string k_PreviewFolderPrefix = "Assets/GraphicsTests/0x_Base/0003_Preview";
 
-        static TestCaseData[] s_TestCaseData = {
+        static TestCaseData[] s_TestCaseData =
+        {
+            new TestCaseData(new MaterialFactory($"TestMaterial-Built-In"), null)
+                .SetName($"Preview generation for Material Built-In")
+                .Returns(null),
             new TestCaseData(new MaterialFactory($"TestMaterial-{nameof(UniversalRenderPipelineAsset)}"), typeof(UniversalRenderPipelineAsset))
                 .SetName($"Preview generation for Material {nameof(UniversalRenderPipelineAsset)}")
                 .Returns(null),
             new TestCaseData(new MaterialFactory($"TestMaterial-{nameof(HDRenderPipelineAsset)}"), typeof(HDRenderPipelineAsset))
                 .SetName($"Preview generation for Material {nameof(HDRenderPipelineAsset)}")
                 .Returns(null)
-                .Ignore("HD Render Pipeline thumbnails doesn't render properly when rendering requested from C++ worker in case thumbnail was scheduled to be rendered later. Possible cause that burstified Job required for proper feature flags determination is not running.")
+                .Ignore(
+                    "UUM-33619: HD Render Pipeline thumbnails doesn't render properly when rendering requested from C++ worker in case thumbnail was scheduled to be rendered later. Possible cause that burstified Job required for proper feature flags determination is not running."),
+            new TestCaseData(new GameObjectFactory($"Hammer.fbx"), null)
+            .SetName($"Preview generation for Model Built-In")
+            .Returns(null),
+            new TestCaseData(new GameObjectFactory($"Hammer.fbx"), typeof(UniversalRenderPipelineAsset))
+                .SetName($"Preview generation for Model {nameof(UniversalRenderPipelineAsset)}")
+                .Returns(null),
+            new TestCaseData(new GameObjectFactory($"Hammer.prefab"), null)
+                .SetName($"Preview generation for Prefab Built-In")
+                .Returns(null),
+            new TestCaseData(new GameObjectFactory($"Hammer.prefab"), typeof(UniversalRenderPipelineAsset))
+                .SetName($"Preview generation for Prefab {nameof(UniversalRenderPipelineAsset)}")
+                .Returns(null),
         };
 
         string m_CreatedObjectPath;
+        AssetFactoryResultStatus m_AssetFactoryAssetFactoryResultStatus;
 
         [UnityTest]
         [TestCaseSource(nameof(s_TestCaseData))]
@@ -40,11 +57,13 @@ namespace Preview
                 yield return null;
 
                 //Arrange
-                var testObject = ProduceNewObject(objectFactory, out m_CreatedObjectPath);
+                var testObject = ProduceNewObject(objectFactory, renderPipelineAssetType, out m_CreatedObjectPath);
                 var loadIcon = LoadIcon(TestContext.CurrentContext.Test.Name);
 
                 //Act
-                var tex = CreatePreviewForAsset(testObject, m_CreatedObjectPath, 128, 128);
+
+                yield return WaitForShadersToCompile(testObject, m_CreatedObjectPath);
+                var tex = AssetPreviewUpdater.CreatePreviewForAsset(testObject, null, m_CreatedObjectPath);
 
                 //Assert
                 ImageAssert.AreEqual(loadIcon, tex, new ImageComparisonSettings
@@ -63,10 +82,11 @@ namespace Preview
         {
             yield return null;
 
-            if(!string.IsNullOrEmpty(m_CreatedObjectPath) && AssetDatabase.AssetPathExists(m_CreatedObjectPath))
+            if (m_AssetFactoryAssetFactoryResultStatus == AssetFactoryResultStatus.Created && !string.IsNullOrEmpty(m_CreatedObjectPath) && AssetDatabase.AssetPathExists(m_CreatedObjectPath))
+            {
                 AssetDatabase.DeleteAsset(m_CreatedObjectPath);
-
-            yield return null;
+                AssetDatabase.Refresh();
+            }
         }
 
         Texture2D LoadIcon(string name)
@@ -76,27 +96,38 @@ namespace Preview
             return loadIcon;
         }
 
-        Object ProduceNewObject(AssetFactory objectFactory, out string path)
+        Object ProduceNewObject(AssetFactory objectFactory, Type renderPipelineType, out string path)
         {
-            var producedObject = objectFactory.GetObjectWithPath(k_PreviewFolderPrefix);
-            path = producedObject.Item2;
-            var testObject = producedObject.Item1;
-            AssetDatabase.CreateAsset(testObject, path);
-            AssetDatabase.Refresh(ImportAssetOptions.ForceUpdate);
+            m_AssetFactoryAssetFactoryResultStatus = objectFactory.GetObjectWithPath(k_PreviewFolderPrefix, renderPipelineType, out var testObject, out path);
             return testObject;
         }
 
-        public Texture2D CreatePreviewForAsset(Object testObject, string path, int width, int height)
+        IEnumerator WaitForShadersToCompile(Object testObject, string path, float maxTime = 300)
         {
-            var assetPreviewUpdaterType = Type.GetType("UnityEditor.AssetPreviewUpdater, UnityEditor.CoreModule, Version=0.0.0.0, Culture=neutral, PublicKeyToken=null");
-            var method = assetPreviewUpdaterType.GetMethod("CreatePreview", BindingFlags.Static | BindingFlags.Public);
-            return method.Invoke(null, new object[] { testObject, null, path, width, height }) as Texture2D;
+            AssetPreviewUpdater.CreatePreviewForAsset(testObject, null, path);
+            yield return null;
+            yield return WaitForChange(() => !ShaderUtil.anythingCompiling, maxTime);
         }
+
+        static IEnumerator WaitForChange(Func<bool> changeCheck, double maxWaitTime = 1)
+        {
+            var startTime = EditorApplication.timeSinceStartup;
+            while (!changeCheck.Invoke() && startTime + maxWaitTime > EditorApplication.timeSinceStartup)
+            {
+                yield return null;
+            }
+        }
+    }
+
+    public enum AssetFactoryResultStatus
+    {
+        Created,
+        Loaded
     }
 
     public abstract class AssetFactory
     {
-        public abstract (Object, string) GetObjectWithPath(string prefix);
+        public abstract AssetFactoryResultStatus GetObjectWithPath(string prefix, Type renderPipelineType, out Object newObject, out string path);
 
         public string name;
 
@@ -108,24 +139,52 @@ namespace Preview
 
     public class MaterialFactory : AssetFactory
     {
+
         public MaterialFactory(string name) : base(name)
         {
         }
 
-        public override (Object, string) GetObjectWithPath(string prefix)
+        public override AssetFactoryResultStatus GetObjectWithPath(string prefix, Type renderPipelineType, out Object newObject, out string path)
         {
-            string shaderName = "";
-            if(GraphicsSettings.currentRenderPipelineAssetType == typeof(UniversalRenderPipelineAsset))
-                shaderName = "Universal Render Pipeline/Lit";
-            else if(GraphicsSettings.currentRenderPipelineAssetType == typeof(HDRenderPipelineAsset))
-                shaderName = "HDRP/Lit";
+            string shaderName;
+            if (renderPipelineType != null)
+            {
+                var asset = RenderPipelineUtils.LoadAsset(renderPipelineType);
+                shaderName = asset.defaultShader.name;
+            }
+            else
+                shaderName = "Standard";
 
-            Assert.That(shaderName, Is.Not.Empty, $"Can not test Material Preview outside of predefined list of SRPs. Current RenderPipelineType is {GraphicsSettings.currentRenderPipelineAssetType?.FullName}");
+            Assert.That(shaderName, Is.Not.Empty,
+                $"Can not test Material Preview outside of predefined list of SRPs. Current RenderPipelineType is {GraphicsSettings.currentRenderPipelineAssetType?.FullName}");
 
             var shader = Shader.Find(shaderName);
-            var newMaterial = new Material(shader);
-            var materialPath = $"{prefix}/{name}.mat";
-            return (newMaterial, materialPath);
+            newObject = new Material(shader);
+            path = $"{prefix}/{name}.mat";
+
+            AssetDatabase.CreateAsset(newObject, path);
+            AssetDatabase.Refresh(ImportAssetOptions.ForceUpdate);
+
+            return AssetFactoryResultStatus.Created;
+        }
+    }
+
+    public class GameObjectFactory : AssetFactory
+    {
+        public GameObjectFactory(string name) : base(name)
+        {
+        }
+
+        public override AssetFactoryResultStatus GetObjectWithPath(string prefix, Type renderPipelineType, out Object newObject, out string path)
+        {
+            path = $"{AssetPreviewTests.k_PreviewFolderPrefix}/{name}";
+            AssetDatabase.ImportAsset(path, ImportAssetOptions.ForceSynchronousImport | ImportAssetOptions.ForceUpdate);
+
+            newObject = AssetDatabase.LoadAssetAtPath<GameObject>(path);
+
+            Assert.That(newObject, Is.Not.Null, $"Couldn't load a {name} by this {path} for testing.");
+
+            return AssetFactoryResultStatus.Loaded;
         }
     }
 }
